+++ packages/common/src/compiler/translator.mock.ts
@@ -423,7 +423,7 @@ export const LIGHTDASH_TABLE_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS_SNOWFLAKE: Om
             fieldType: FieldType.DIMENSION,
             description: undefined,
             type: DimensionType.DATE,
-            sql: "DATE_TRUNC('DAY', ${TABLE}.user_created)",
+            sql: "DATE_TRUNC('DAY', CONVERT_TIMEZONE('UTC', ${TABLE}.user_created))",
             name: 'user_created_day',
             table: MODEL_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS.name,
             tableLabel: 'My table',
@@ -441,7 +441,7 @@ export const LIGHTDASH_TABLE_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS_SNOWFLAKE: Om
             fieldType: FieldType.DIMENSION,
             description: undefined,
             type: DimensionType.DATE,
-            sql: "DATE_TRUNC('WEEK', ${TABLE}.user_created)",
+            sql: "DATE_TRUNC('WEEK', CONVERT_TIMEZONE('UTC', ${TABLE}.user_created))",
             name: 'user_created_week',
 
             table: MODEL_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS.name,
@@ -460,7 +460,7 @@ export const LIGHTDASH_TABLE_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS_SNOWFLAKE: Om
             fieldType: FieldType.DIMENSION,
             description: undefined,
             type: DimensionType.DATE,
-            sql: "DATE_TRUNC('MONTH', ${TABLE}.user_created)",
+            sql: "DATE_TRUNC('MONTH', CONVERT_TIMEZONE('UTC', ${TABLE}.user_created))",
             name: 'user_created_month',
 
             table: MODEL_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS.name,
@@ -479,7 +479,7 @@ export const LIGHTDASH_TABLE_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS_SNOWFLAKE: Om
             fieldType: FieldType.DIMENSION,
             description: undefined,
             type: DimensionType.DATE,
-            sql: "DATE_TRUNC('YEAR', ${TABLE}.user_created)",
+            sql: "DATE_TRUNC('YEAR', CONVERT_TIMEZONE('UTC', ${TABLE}.user_created))",
             name: 'user_created_year',
 
             table: MODEL_WITH_DEFAULT_TIME_INTERVAL_DIMENSIONS.name,